Change data type

+++ TimeConversion/TimeConversion/Converter/Converter.swift
@@ -10,7 +10,7 @@ import Foundation
 class Converter {
     private init() {}
 
-    static func convert(value: Int, from: Unit, to: Unit) -> Int {
+    static func convert(value: Double, from: Unit, to: Unit) -> Double {
         if from == to {
             return value
         }
@@ -42,7 +42,7 @@ class Converter {
         return 0
     }
 
-    private static func convertBetweenSecondsAndMinutes(value: Int, from: Unit, to: Unit) -> Int {
+    private static func convertBetweenSecondsAndMinutes(value: Double, from: Unit, to: Unit) -> Double {
         if from == .seconds {
             return value / 60
         }
@@ -50,7 +50,7 @@ class Converter {
         return value * 60
     }
 
-    private static func convertBetweenSecondsAndHours(value: Int, from: Unit, to: Unit) -> Int {
+    private static func convertBetweenSecondsAndHours(value: Double, from: Unit, to: Unit) -> Double {
         if from == .seconds {
             return value / 60 / 60
         }
@@ -58,7 +58,7 @@ class Converter {
         return value * 60 * 60
     }
 
-    private static func convertBetweenSecondsAndDays(value: Int, from: Unit, to: Unit) -> Int {
+    private static func convertBetweenSecondsAndDays(value: Double, from: Unit, to: Unit) -> Double {
         if from == .seconds {
             return value / 60 / 60 / 24
         }
@@ -66,7 +66,7 @@ class Converter {
         return value * 60 * 60 * 24
     }
 
-    private static func convertBetweenMinutesAndHours(value: Int, from: Unit, to: Unit) -> Int {
+    private static func convertBetweenMinutesAndHours(value: Double, from: Unit, to: Unit) -> Double {
         if from == .minutes {
             return value / 60
         }
@@ -74,7 +74,7 @@ class Converter {
         return value * 60
     }
 
-    private static func convertBetweenMinutesAndDays(value: Int, from: Unit, to: Unit) -> Int {
+    private static func convertBetweenMinutesAndDays(value: Double, from: Unit, to: Unit) -> Double {
         if from == .minutes {
             return value / 60 / 24
         }
@@ -82,7 +82,7 @@ class Converter {
         return value * 60 * 24
     }
 
-    private static func convertBetweenHoursAndDays(value: Int, from: Unit, to: Unit) -> Int {
+    private static func convertBetweenHoursAndDays(value: Double, from: Unit, to: Unit) -> Double {
         if from == .hours {
             return value / 24
         }
